P3 increment 1: shaft-capture-proxy module scaffold + shared file-permission utility
New shaft-capture-proxy Maven module (Phase P3 of #3346 -- mobile API
recording), kept separate from shaft-capture/shaft-engine so the
LittleProxy/Netty/Bouncy Castle dependencies it needs for an embedded
MITM proxy stay off the classpath of consumers that never enable mobile
API capture. Verified io.github.littleproxy:littleproxy:2.4.2 (the
maintained fork; org.littleshoot:littleproxy is unmaintained) resolves
cleanly against real Maven Central coordinates and shades its own Netty
dependency with no conflict against Selenium's shaded grpc-netty.
Bouncy Castle (bcprov/bcpkix-jdk18on) was already a managed dependency
in the root POM at the version this module needs.

Extracted OwnerOnlyFilePermissions (POSIX 0600 / Windows owner-only ACL
fallback) out of SecretHeaderReplacer into a shared, independently
tested utility in shaft-capture, since the new CA private key this
module will generate needs the exact same protection P1's .secrets.json
already had -- one tested implementation of a security-sensitive
operation instead of two copies. Refactor verified against the existing
SecretHeaderReplacerTest (14/14, no regression) plus a new dedicated
OwnerOnlyFilePermissionsTest (2/2).
