+++ shaft-bom/pom.xml
@@ -32,6 +32,11 @@
             </dependency>
             <dependency>
                 <groupId>${project.groupId}</groupId>
+                <artifactId>shaft-capture-proxy</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${project.groupId}</groupId>
                 <artifactId>shaft-doctor</artifactId>
                 <version>${project.version}</version>
             </dependency>
