+++ shaft-capture-proxy/src/main/java/com/shaft/capture/proxy/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Embedded MITM HTTP(S) proxy for recording native mobile API traffic into SHAFT Capture
+ * sessions.
+ */
+package com.shaft.capture.proxy;
